src/indicator-service/requirements.txt
pandas>=2.3.2
pandas-ta==0.4.71b0
confluent-kafka==2.8.0
clickhouse-connect==0.8.14
fastapi==0.115.12
uvicorn==0.34.2
python-dotenv==1.1.0
structlog==25.1.0
pytz==2025.2
